refactor: type 관리 변경으로 인한 import 추가
