+++ client/src/components/Modal/SetActionModalContent/SetActionModalContent.tsx
@@ -1,3 +1,5 @@
+import type {InOutType} from 'types/serviceType';
+
 import {useState} from 'react';
 import {BottomSheet, Switch} from 'haengdong-design';
